Split up the with items unit tests for the conductor
For clarity, split up cancel, pause, and resume related unit tests to separate test modules.
